Added Multi-head attention
in the first version, the multi-head attention is a generalization of the
scaled dot-product attention, but the attention API could be generalized more
to allow the multi-head attention to be a wrapper for any kind of attention
mechanism.

TODOs include user-friendly control of parameters (decoder.rnn_size (query
size) should match the encoder hidden state sizes (i.e. values and key sizes))

Theoretically, only the query vector should match the "key"-encoder hidden
state size. But that is also specific for the dot-product attention.